k-distribution/kast: make prettyPrintKast closer to actual `kast --output pretty` output
