+++ k-distribution/src/main/scripts/lib/pyk/kast.py
@@ -3,6 +3,19 @@
 import sys
 import json
 
+def _notif(msg):
+    sys.stderr.write('\n')
+    sys.stderr.write(msg + '\n')
+    sys.stderr.write(''.join(['=' for c in msg]) + '\n')
+    sys.stderr.flush()
+
+def _warning(msg):
+    _notif('[WARNING] ' + msg)
+
+def _fatal(msg, code = 1):
+    _notif('[FATAL] ' + msg)
+    sys.exit(code)
+
 def combineDicts(*dicts):
     if len(dicts) == 0:
         return {}
@@ -117,7 +130,7 @@ def binOpStr(symbol):
     return (lambda a1, a2: a1 + " " + symbol + " " + a2)
 
 def appliedLabelStr(symbol):
-    return (lambda *args: symbol + "( " + " , ".join(args) + " )")
+    return (lambda *args: symbol + " ( " + " , ".join(args) + " )")
 
 def constLabel(symbol):
     return (lambda: symbol)
@@ -175,10 +188,10 @@ INT_expressions = { "_+Int_"   : paren(binOpStr("+Int"))
 
 MAP_expressions = { "Map:update"  : paren(underbarUnparsing("_[_<-_]"))
                   , "Map:lookup"  : paren(underbarUnparsing("_[_]"))
-                  , "_Map_"       : underbarUnparsing("_\n_")
-                  , "_|->_"       : paren(underbarUnparsing("_|->_"))
+                  , "_Map_"       : lambda m1, m2: m1 + "\n" + m2
+                  , "_|->_"       : underbarUnparsing("_|->_")
                   , "_[_<-undef]" : paren(underbarUnparsing("_[_ <- undef ]"))
-                  , ".Map"        : constLabel(".Map")
+                  , ".Map"        : constLabel("")
                   }
 
 LIST_expressions = { "_List_"   : (lambda a1, a2: a1 + " " + a2)
@@ -202,17 +215,19 @@ def prettyPrintKast(kast, symbolTable):
     if isKApply(kast):
         label = kast["label"]
         args  = kast["args"]
+        unparsedArgs = [ prettyPrintKast(arg, symbolTable) for arg in args ]
         if isCellKLabel(label):
-            cellLines = "\n".join([ prettyPrintKast(arg, symbolTable) for arg in args ]).split("\n")
-            cellStr   = "\n  ".join(cellLines)
-            if len(cellLines) == 1:
-                cellStr = label + " "    + cellStr + " </"  + label[1:]
-            else:
-                cellStr = label + "\n  " + cellStr + "\n</" + label[1:]
-            return cellStr
-        if label in symbolTable:
-            newArgs = [prettyPrintKast(arg, symbolTable) for arg in args]
-            return symbolTable[kast["label"]](*newArgs)
+            cellLines = "\n".join(unparsedArgs).rstrip().split("\n")
+            cellStr   = label + "\n  " + "\n  ".join(cellLines) + "\n</" + label[1:]
+            return cellStr.rstrip()
+        unparser = appliedLabelStr(label) if label not in symbolTable else symbolTable[label]
+        return unparser(*unparsedArgs)
+    if isKSequence(kast):
+        unparsedItems = [ prettyPrintKast(item, symbolTable) for item in kast['items'] ]
+        unparsedKSequence = "\n~> ".join(unparsedItems)
+        if len(unparsedItems) > 1:
+            unparsedKSequence = "    " + unparsedKSequence
+        return unparsedKSequence
     if isKRule(kast):
         body     = "\n     ".join(prettyPrintKast(kast["rule"], symbolTable).split("\n"))
         ruleStr = "rule " + body
@@ -254,7 +269,7 @@ def prettyPrintKast(kast, symbolTable):
         return requires + "\n\n" + modules
 
     print()
-    warning("Error turning to kast!")
+    _warning("Error unparsing kast!")
     print(kast)
-    fatal("Error unparsing!")
+    _fatal("Error unparsing!")
